requirements.txt
evol~=0.5.3
pillow~=9.0.0
numpy~=1.21.4
configparser~=5.2.0
matplotlib~=3.5.2